requirements.txt
-r requirements_dev.txt
amqp==1.4.8
anyjson==0.3.3
apipkg==1.4
argh==0.26.1
billiard==3.3.0.22
celery==3.1.19
colorama==0.3.5
decorator==4.0.6
Django==1.9
django-bower==5.0.4
django-braces==1.8.1
django-crispy-forms==1.5.2
django-enumfield==1.2.1
django-enumfields==0.7.4
django-extensions==1.6.1
django-redis==4.3.0
djangorestframework==3.3.2
docker-py==1.6.0
docopt==0.6.2
drf-enum-field==0.9.1
execnet==1.4.1
factory-boy==2.6.0
fake-factory==0.5.3
freezegun==0.3.5
gunicorn==19.4.1
ipython==4.0.1
ipython-genutils==0.1.0
jedi==0.9.0
kombu==3.0.32
path.py==8.1.2
pathtools==0.1.2
pexpect==4.0.1
pickleshare==0.5
prompt-toolkit==0.54
psycopg2==2.6.1
ptpython==0.25
ptyprocess==0.5
py==1.4.31
Pygments==2.0.2
python-dateutil==2.4.2
python-redis-lock==2.3.0
pytz==2015.7
PyYAML==3.11
redis==2.10.5
requests==2.9.1
simplegeneric==0.8.1
six==1.10.0
termcolor==1.1.0
traitlets==4.0.0
watchdog==0.8.3
wcwidth==0.1.5
websocket-client==0.34.0
